Split into files
Redesign of code structure + basic tests
